Add Rotate for Minimum Width extension
Helpful when trying to nest objects, this rotates all selected objects
to minimize their width.

This first-pass implementation is pretty basic: it tries rotation 90
degrees in either direction in half-degree increments, whereas a more
efficient approach would narrow in on the ideal angle, binary-search
style. This is pretty slow, but works for my purposes for now.